+++ assets/js/public/job-application.js
@@ -107,6 +107,9 @@ jQuery(document).ready(function($) {
 				})
 				.always(function() {
 					$submitBtn.prop('disabled', false).val(submitBtnText).removeClass('awsm-application-submit-btn-disabled');
+					if (typeof grecaptcha !== 'undefined') {
+						grecaptcha.reset();
+					}
 				});
 		}
 	};